Created class converter, for simplify class OvalNode
